creating the delete project template and small form to handle second validation before deleting the project

+++ app/admin/views.py
@@ -67,17 +67,24 @@ def new_project():
 def edit_project(id):
     project = Project.query.get_or_404(id)
     if request.method == 'POST':
-        project.title = request.form.get('title')
-        project.small_description = request.form.get('small_description')
-        project.description = request.form.get('description')
-        project.image_file = request.form.get('image_file')
-        project.github_link = request.form.get('github_link')
-        project.project_link = request.form.get('project_link')
-        project.slug = slugify(project.title)
-    
+        project.title = request.form['title']
+        project.small_description = request.form['small-description']
+        project.description = request.form['description']
+        project.github_link = request.form['github-link']
+        project.project_link = request.form.get('project-link', '')
+
+        if 'image' in request.files:
+            file = request.files['image']
+            if file.filename != '':
+                filename = secure_filename(file.filename)
+                file_path = os.path.join(current_app.config['UPLOAD_FOLDER'], filename)
+                file.save(file_path)
+                project.image_file = filename
+
         db.session.commit()
-        flash('Project updated successfulfy', 'success')
+        flash('Project updated successfully!')
         return redirect(url_for('admin.dashboard'))
+
     return render_template('admin/edit_project.html', project=project)
 
 
@@ -85,7 +92,11 @@ def edit_project(id):
 @admin.route('/admin/project/delete/<int:id>', methods=['GET', 'POST'])
 def delete_project(id):
     project = Project.query.get_or_404(id)
-    db.session.delete(project)
-    db.session.commit()
-    flash('Project deleted successfully', 'success')
-    return redirect(url_for('admin.dashboard'))
+    if request.method=='POST':
+        if request.form.get('response') == 'Yes':
+            db.session.delete(project)
+            db.session.commit()
+            flash('Project deleted successfully', 'success')
+            return redirect(url_for('admin.dashboard'))
+        return redirect(url_for('admin.dashboard'))
+    return render_template('admin/delete_project.html', project=project)
